Table: Support `obj[key]`
